pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "teamsservicebell"
version = "0.0.1"
authors = [
  {name = "Eric Lewis", email = "thedreadpirateeric@gmail.com"},
]
maintainers = [
  {name = "Eric Lewis", email = "thedreadpirateeric@gmail.com"},
]
description = "Talks to teams at the press of a button"
readme = "README.md"
requires-python = ">= 3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

dependencies = [
  "pymsteams",
  "pyyaml"
]

[project.urls]
Homepage = "https://github.com/thedreadpirateeric/teams-service-bell"
Issues = "https://github.com/thedreadpirateeric/teams-service-bell/issues"


[[tool.mypy.overrides]]
module = [
  "pymsteams.*",
  "pyyaml.*"
]
follow_untyped_imports = true
